+++ gnark_backend_ffi/backend/plonk/black_box_functions.go
@@ -1,0 +1,40 @@
+package plonk_backend
+
+// AES black box function call is not handled
+func AES() {}
+
+// AND black box function call is not handled
+func AND() {}
+
+// XOR black box function call is not handled
+func XOR() {}
+
+// RANGE black box function call is not handled
+func Range() {}
+
+// SHA256 black box function call is not handled
+func SHA256() {}
+
+// Blake2s black box function call is not handled
+func Blake2s() {}
+
+// MerkleMembership black box function call is not handled
+func MerkleMembership() {}
+
+// SchnorrVerify black box function call is not handled
+func SchnorrVerify() {}
+
+// Pedersen black box function call is not handled
+func Pedersen() {}
+
+// HashToField128Security black box function call is not handled
+func HashToField128Security() {}
+
+// EcdsaSecp256k1 black box function call is not handled
+func EcdsaSecp256k1() {}
+
+// FixedBaseScalarMul black box function call is not handled
+func FixedBaseScalarMul() {}
+
+// Keccak256 black box function call is not handled
+func Keccak256() {}
